requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
apispec==6.4.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
flask-restx==1.3.0
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
frozenlist==1.4.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
mistune==3.0.2
mongoengine==0.27.0
multidict==6.0.5
ordered-set==4.1.0
packaging==23.2
py-mon==2.0.3
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
twilio==8.12.0
typing_extensions==4.9.0
urllib3==2.2.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
